requirements.txt
# Core AI and ML dependencies
langchain==0.1.12
langchain-openai==0.1.0
openai==1.12.0
chromadb==0.4.22

# Web framework and API
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.29.0

# Code analysis and parsing
astunparse==1.6.3
black==23.11.0
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Testing and coverage
pytest==7.4.3
pytest-cov==4.1.0
coverage==7.3.2

# Data processing and utilities
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
pydantic==2.5.0

# HTTP and networking
requests==2.31.0
httpx==0.25.2

# File handling and utilities
pathlib2==2.3.7
click==8.1.7
rich==13.7.0
typer==0.9.0

# Development and debugging
ipython==8.18.1
jupyter==1.0.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Security and validation
cryptography==41.0.8
python-multipart==0.0.6

# Additional utilities
tqdm==4.66.1
colorama==0.4.6 